Excl. leading dir: directory content move func
Add function moveFilesExcludingLeadingDir in Unzipper.
Function moves all files and folders from one directory to another, but
if the source folder only contains a singe directory, only the content
of tha single directory will be moved, not the directory itself.

This is a part of a bigger update to add the feature of excluding a
leading directory in an archive when extracting.